+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix spell check to use chosen application language where possible
 - Fix stack error preventing use of rnd() within "if" statements
+- Fix 'wait for input' text code responsiveness when using slower text speeds
 
 ## [4.0.1] - 2024-07-22